+++ CHANGELOG.md
@@ -20,6 +20,7 @@ New features
 * New syntax for `for` statement to include counter variable.
 * An integer value can now be indexed to get/set a single bit.
 * The `bits` method of an integer can be used to iterate through its bits.
+* New `$bool$`, `$int$`, `$float$` and `$string$` expression types for custom syntax.
 
 
 Version 0.20.2